Move manage integrations button in to stand-alone component
